extract xrp code to gem_xrp crate
