--- crates/gem_chain_rpc/src/xrp/mod.rs
@@ -1,4 +1,0 @@
-pub mod client;
-pub mod mapper;
-pub mod model;
-pub mod provider;
