+++ crates/gem_xrp/src/rpc/mod.rs
@@ -1,0 +1,6 @@
+mod client;
+pub mod mapper;
+pub mod provider;
+
+pub use self::client::XRPClient;
+pub use self::provider::XRPProvider;
