+++ crates/settings_chain/Cargo.toml
@@ -12,3 +12,4 @@ primitives = { path = "../primitives" }
 gem_chain_rpc = { path = "../gem_chain_rpc" }
 gem_solana = { path = "../gem_solana" }
 settings = { path = "../settings" }
+gem_xrp = { path = "../gem_xrp", features = ["rpc"] }
